refactor: change structure of requests page
