+++ src/pages/requests.js
@@ -1,0 +1,8 @@
+import React from 'react';
+import RequestsTable from '../components/requests-table';
+
+const Requests = () => {
+  return <RequestsTable />;
+};
+
+export default Requests;
